pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-security-guardrails"
version = "0.1.0"
description = "Security guardrails, input/output validation, and governance controls for LLM applications, agents, and RAG systems"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
  { name = "Hiago Kin Levi" }
]
keywords = ["llm", "security", "guardrails", "ai-safety", "prompt-injection", "rag"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  # Data validation and settings management
  "pydantic>=2.6",
  "python-dotenv>=1.0",

  # Web framework and server
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",

  # CLI
  "click>=8.1",

  # Structured logging
  "structlog>=24.1",

  # LLM client and token counting
  "openai>=1.30",
  "tiktoken>=0.7",

  # Template rendering (for reports and prompts)
  "jinja2>=3.1",

  # YAML policy parsing (stdlib PyYAML)
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "httpx>=0.27",         # For testing FastAPI endpoints
  "ruff>=0.4",
  "mypy>=1.10",
  "types-PyYAML",
]

[project.scripts]
# CLI entry point
k1n-guardrails = "guardrails.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["guardrails", "middleware", "schemas"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S"]
ignore = [
  "S101",   # allow assert statements in tests
]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
